requirements.txt
chromadb>=0.5.3
langchain_core>=0.3.75
langchain_groq>=0.3.7
numpy>=2.3.3
PyPDF2>=3.0.1
python-dotenv>=1.1.1
Requests>=2.32.5
sentence_transformers>=5.0.0
streamlit>=1.45.1
pysqlite3-binary>=0.5.4